+++ services/reviews.go
@@ -70,9 +70,8 @@ func (s *Reviews) GetReviews(ctx context.Context, req *reviews.Product) (*review
 	session := s.MongoSession.Copy()
 	defer session.Close()
 	c := session.DB("reviews-db").C("reviews")
-
 	var result []DB_Review;
-	err := c.Find(&bson.M{"ProductID": int(productID)}).Limit(100).All(&result)
+	err := c.Find(&bson.M{"ProductID": int(productID)}).All(&result)
 	if err != nil {
 		log.Fatalf("Try to find product id [%v], err = %v", productID, err.Error())
 	}
@@ -91,25 +90,11 @@ func (s *Reviews) GetReviews(ctx context.Context, req *reviews.Product) (*review
 	}
 	log.Printf("Got result num %v, id = %v", len(res.Review), productID)
 
-	// review1 := reviews.Review{
-	// 	ProductID: productID,
-	// 	Reviewer:  "reviewer1",
-	// 	Text:      "An extremely entertaining play by Shakespeare. The slapstick humour is refreshing!",
-	// }
-
-	// review2 := reviews.Review{
-	// 	ProductID: productID,
-	// 	Reviewer:  "reviewer2",
-	// 	Text:      "Absolutely fun and entertaining. The play lacks thematic depth when compared to other plays by Shakespeare.",
-	// }
-
 	version := os.Getenv("REVIEWS_VERSION")
 
-	//res.Review = append(res.Review, &review1)
-	//res.Review = append(res.Review, &review2)
 
 	if version != "v1" {
-		log.Printf("Sending request to ratings service")
+		log.Printf("Sending request to ratings service, id = %v", productID)
 		ratingsRes, err := s.ratingsClient.GetRatings(ctx, &ratings.Product{
 			Id: int32(productID),
 		})
